[REF] Major refactoring of package
To conform to expected format of traditional Python packages. Also made
some minor changes to fix linting, etc.
